Reestructure HTML and styles
The old version of this page was not styled, only the index had some simple style. In this commit i've reestructured the html, mainly at the recipes files, adding classes to be able to style then. So, i've created a recipe-style.css file where i have styled the recipes using flexbox and other properties.